26/requirements.txt
pillow==10.4.0
ttkbootstrap==1.12.2
